requirements.txt
paho-mqtt>=2.0.0
sounddevice>=0.4.6
numpy>=1.24.0
gpiod>=2.0.0
